working output!

+++ cgfs/viewport.py
@@ -1,5 +1,5 @@
 class Viewport:
-    def __init__(self, width: int, height: int, distance: int):
+    def __init__(self, width: float, height: float, distance: float):
         self.width = width
         self.height = height
         self.distance = distance
